+++ src/ymm/ymm.py
@@ -1,0 +1,7 @@
+
+class YMM:
+    def __init__(self, yaml):
+        self.yaml = yaml
+
+    def run(self):
+        return True
